Temp commit - initial version working generating a DocFx on .NET 8, Powershell 5.1 and docfx 2.60.1 - running without docfx.console.exe - clicking top bar now goes -1 in history, if no history goes to root - install/manual link not generated if no install.md nor manual folder exists at root - allows empty/null values for footer giturl, websiteurl and nugeturl

+++ scripts/Powershell/06-clean-output-dir.ps1
@@ -1,7 +1,13 @@
 try {
     if ($outputFolderFullPath -ne $null -and $outputFolderFullPath -ne "") {
-        Remove-Item -Recurse -Force $outputFolderFullPath -ErrorAction SilentlyContinue
-        Out ("Cleaned output directory: " + $outputFolderFullPath)
+
+        if ($outputFolderFullPath.Length -gt 4) {
+            Remove-Item -Recurse -Force $outputFolderFullPath -ErrorAction SilentlyContinue
+            Out ("Cleaned output directory: " + $outputFolderFullPath)
+        }
+        else {
+            Err "Cannot clear output dir, it is too short, minimum 4 chars"
+        }
     }
 }
 catch {
